feature(lowpower): add low voltage protection
1. add adc read with calibration
2. enter deep-sleep and wake up every 3 seconds when supply voltage is lower than 2.8V

+++ examples/lowpower_evb/components/lowpower_evb_peripherals/include/lowpower_evb_adc.h
@@ -28,9 +28,9 @@ void adc_init();
   * @brief get supply voltage.
   *
   * @return 
-  *     - the value of sulpply_voltage. uint:V
+  *     - the value of sulpply_voltage. uint:mV
   */
-float adc_get_supply_voltage();
+uint32_t adc_get_supply_voltage();
 
 #ifdef __cplusplus
 }
